requirements_db.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
alembic==1.12.0
python-dotenv==1.0.0